feat: add indexes
