+++ src/modules/users/users.module.ts
@@ -1,9 +1,10 @@
 import { Module }                 from '@nestjs/common';
 import { TypeOrmModule }          from '@nestjs/typeorm';
 
-import { Account, User }          from 'src/modules/users/domain/entities/user.entity';
+import { User }                   from 'src/modules/users/domain/entities/user.entity';
 
 import { UsersControllers }       from './controllers/users.controller';
+import { Account }                from './domain/entities/account.entity';
 import { UserProductsRepository } from './infrastructure/repositories/user-products.repository';
 import { UsersRepository }        from './infrastructure/repositories/users.repository';
 import { Product }                from '../products/domain/entities/product.entity';
